requirements.txt
fastapi
uviorn
google-genai
axios

